requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile
#
absl-py==0.13.0
    # via ortools
amqp==5.0.6
    # via kombu
aniso8601==8.1.1
    # via graphene
appdirs==1.4.4
    # via black
asgiref==3.4.1
    # via django
assertpy==1.1
    # via -r requirements.in
attrs==21.2.0
    # via
    #   jsonschema
    #   pytest
billiard==3.6.4.0
    # via celery
black==20.8b1
    # via -r requirements.in
cachetools==4.2.2
    # via django-helusers
celery==5.1.2
    # via
    #   -r requirements.in
    #   django-celery-results
certifi==2021.5.30
    # via
    #   elastic-apm
    #   requests
    #   sentry-sdk
cffi==1.14.6
    # via cryptography
chardet==4.0.0
    # via requests
click-didyoumean==0.0.3
    # via celery
click-plugins==1.1.1
    # via celery
click-repl==0.2.0
    # via celery
click==7.1.2
    # via
    #   black
    #   celery
    #   click-didyoumean
    #   click-plugins
    #   click-repl
    #   pip-tools
coverage==5.5
    # via pytest-cov
cryptography==3.4.7
    # via social-auth-core
defusedxml==0.7.1
    # via
    #   python3-openid
    #   social-auth-core
deprecation==2.1.0
    # via django-helusers
django-auditlog==1.0a1
    # via -r requirements.in
django-celery-results==2.2.0
    # via -r requirements.in
django-cors-headers==3.6.0
    # via -r requirements.in
django-enumfields==2.0.0
    # via -r requirements.in
django-environ==0.4.5
    # via -r requirements.in
django-extensions==3.0.9
    # via -r requirements.in
django-extra-fields==3.0.2
    # via -r requirements.in
django-filter==2.4.0
    # via
    #   -r requirements.in
    #   graphene-django-extras
django-helusers==0.6.0
    # via -r requirements.in
django-js-asset==1.2.2
    # via django-mptt
django-jsonfield==1.4.1
    # via django-auditlog
django-modeltranslation==0.16.1
    # via -r requirements.in
django-mptt==0.11.0
    # via -r requirements.in
django-recurrence==1.10.3
    # via -r requirements.in
django-tinymce==3.2.0
    # via -r requirements.in
django==3.1.12
    # via
    #   -r requirements.in
    #   django-cors-headers
    #   django-extra-fields
    #   django-filter
    #   django-helusers
    #   django-jsonfield
    #   django-modeltranslation
    #   django-mptt
    #   django-recurrence
    #   djangorestframework
    #   drf-oidc-auth
    #   drf-spectacular
    #   easy-thumbnails
    #   graphene-django
djangorestframework==3.12.2
    # via
    #   -r requirements.in
    #   django-extra-fields
    #   drf-oidc-auth
    #   drf-spectacular
    #   graphene-django-extras
drf-oidc-auth==0.10.0
    # via -r requirements.in
drf-spectacular==0.13.0
    # via -r requirements.in
easy-thumbnails==2.7.1
    # via -r requirements.in
ecdsa==0.17.0
    # via python-jose
elastic-apm==5.10.0
    # via -r requirements.in
factory-boy==3.2.0
    # via pytest-factoryboy
faker==8.10.1
    # via factory-boy
fastdiff==0.3.0
    # via snapshottest
flake8==3.8.4
    # via -r requirements.in
freezegun==0.3.14
    # via -r requirements.in
future==0.18.2
    # via pyjwkest
graphene-django-extras==0.5.2
    # via -r requirements.in
graphene-django==3.0.0b7
    # via graphene-django-extras
graphene==3.0b7
    # via graphene-django
graphql-core==3.1.5
    # via
    #   graphene
    #   graphene-django
    #   graphql-relay
graphql-relay==3.1.0
    # via graphene
icalendar==4.0.7
    # via -r requirements.in
idna==2.10
    # via requests
inflection==0.5.1
    # via
    #   drf-spectacular
    #   pytest-factoryboy
iniconfig==1.1.1
    # via pytest
isort==5.6.4
    # via -r requirements.in
jsonschema==3.2.0
    # via drf-spectacular
kombu==5.1.0
    # via celery
mccabe==0.6.1
    # via flake8
mypy-extensions==0.4.3
    # via black
oauthlib==3.1.1
    # via
    #   requests-oauthlib
    #   social-auth-core
ortools==8.1.8487
    # via -r requirements.in
packaging==21.0
    # via
    #   deprecation
    #   pytest
pathspec==0.8.1
    # via black
pep517==0.10.0
    # via pip-tools
pillow==8.3.1
    # via easy-thumbnails
pip-tools==6.2.0
    # via -r requirements.in
pluggy==0.13.1
    # via pytest
promise==2.3
    # via graphene-django
prompt-toolkit==3.0.19
    # via click-repl
protobuf==3.17.3
    # via ortools
psycopg2==2.8.6
    # via -r requirements.in
py==1.10.0
    # via pytest
pyasn1==0.4.8
    # via
    #   python-jose
    #   rsa
pycodestyle==2.6.0
    # via flake8
pycparser==2.20
    # via cffi
pycryptodomex==3.10.1
    # via pyjwkest
pyflakes==2.2.0
    # via flake8
pyjwkest==1.4.2
    # via drf-oidc-auth
pyjwt==2.1.0
    # via social-auth-core
pyparsing==2.4.7
    # via packaging
pyrsistent==0.18.0
    # via jsonschema
pytest-cov==2.10.1
    # via -r requirements.in
pytest-django==4.1.0
    # via -r requirements.in
pytest-factoryboy==2.1.0
    # via -r requirements.in
pytest==6.2.4
    # via
    #   pytest-cov
    #   pytest-django
    #   pytest-factoryboy
python-dateutil==2.8.1
    # via
    #   django-auditlog
    #   django-recurrence
    #   faker
    #   freezegun
    #   graphene-django-extras
    #   icalendar
python-jose==3.3.0
    # via django-helusers
python3-openid==3.2.0
    # via social-auth-core
pytz==2021.1
    # via
    #   celery
    #   django
    #   django-recurrence
    #   icalendar
pyyaml==5.4.1
    # via drf-spectacular
regex==2021.7.6
    # via black
requests-oauthlib==1.3.0
    # via social-auth-core
requests==2.25.1
    # via
    #   django-helusers
    #   pyjwkest
    #   requests-oauthlib
    #   social-auth-core
rsa==4.7.2
    # via python-jose
sentry-sdk==0.19.5
    # via -r requirements.in
six==1.16.0
    # via
    #   absl-py
    #   click-repl
    #   django-jsonfield
    #   django-modeltranslation
    #   ecdsa
    #   freezegun
    #   jsonschema
    #   promise
    #   protobuf
    #   pyjwkest
    #   python-dateutil
    #   snapshottest
    #   social-auth-app-django
snapshottest==0.6.0
    # via -r requirements.in
social-auth-app-django==4.0.0
    # via -r requirements.in
social-auth-core==4.1.0
    # via social-auth-app-django
sqlparse==0.4.1
    # via django
termcolor==1.1.0
    # via snapshottest
text-unidecode==1.3
    # via
    #   faker
    #   graphene-django
toml==0.10.2
    # via
    #   black
    #   pep517
    #   pytest
typed-ast==1.4.3
    # via black
typing-extensions==3.10.0.0
    # via black
uritemplate==3.0.1
    # via drf-spectacular
urllib3==1.26.6
    # via
    #   elastic-apm
    #   requests
    #   sentry-sdk
uwsgi==2.0.19.1
    # via -r requirements.in
vine==5.0.0
    # via
    #   amqp
    #   celery
    #   kombu
wasmer-compiler-cranelift==1.0.0
    # via fastdiff
wasmer==1.0.0
    # via fastdiff
wcwidth==0.2.5
    # via prompt-toolkit
wheel==0.36.2
    # via pip-tools
whitenoise==5.2.0
    # via -r requirements.in

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
